A/B Tests RegistrationStatus
